move TS 11.qnamaker, cleanup, bump dependencies
