+++ samples/javascript_typescript/11.qnamaker/.env
@@ -1,0 +1,2 @@
+botFilePath=../qnamaker.bot
+botFileSecret=
