+++ samples/javascript_typescript/11.qnamaker/README.md
@@ -10,21 +10,13 @@ In this sample, we demonstrate how to use the QnA Maker service to answer questi
     ```bash
     git clone https://github.com/microsoft/botbuilder-samples.git
     ```
-- In a terminal, navigate to javascript_ts\11.qnamaker
+- In a terminal, navigate to samples/javascript_typescript/11.qnamaker
     ```bash
-    cd javascript_ts\11.qnamaker
-    ```
-- Point to the Myget feed 
-    ```bash
-    npm config set registry https://botbuilder.myget.org/F/botbuilder-v4-js-daily/npm/
+    cd samples/javascript_typescript/11.qnamaker
     ```
 - Install modules and start the bot
     ```bash
     npm i & npm start
-    ```
-- To reset registry, you can do
-    ```bash
-    npm config set registry https://registry.npmjs.org/
     ```
 
 # Testing the bot using Bot Framework Emulator
@@ -32,9 +24,9 @@ In this sample, we demonstrate how to use the QnA Maker service to answer questi
 
 - Install the Bot Framework Emulator from [here](https://aka.ms/botframework-emulator)
 
-## Connect to bot using Bot Framework Emulator **V4**
+## Connect to bot using Bot Framework Emulator V4
 - Launch Bot Framework Emulator
-- File -> Open Bot Configuration and navigate to javascript_ts\11.qnamaker folder
+- File -> Open Bot Configuration and navigate to samples/javascript_typescript/11.qnamaker folder
 - Select qnamaker.bot file
 
 ## Prerequisites
@@ -47,6 +39,19 @@ In this sample, we demonstrate how to use the QnA Maker service to answer questi
 QnA Maker enables you to power a question and answer service from your semi-structured content. 
 
 One of the basic requirements in writing your own Bot service is to seed it with questions and answers. In many cases, the questions and answers already exist in content like FAQ URLs/documents, product manuals, etc. With QnA Maker, users can query your application in a natural, conversational manner. QnA Maker uses machine learning to extract relevant question-answer pairs from your content. It also uses powerful matching and ranking algorithms to provide the best possible match between the user query and the questions.
+
+# Deploy this bot to Azure
+You can use the [MSBot](https://github.com/microsoft/botbuilder-tools) Bot Builder CLI tool to clone and configure any services this sample depends on. 
+
+To install all Bot Builder tools - 
+```bash
+npm i -g msbot chatdown ludown qnamaker luis-apis botdispatch luisgen
+```
+
+To clone this bot, run
+```
+msbot clone services -f deploymentScripts/msbotClone -n <BOT-NAME> -l <Azure-location> --subscriptionId <Azure-subscription-id>
+```
 
 # Further reading
 - [About Azure Bot Service](https://docs.microsoft.com/en-us/azure/bot-service/bot-service-overview-introduction?view=azure-bot-service-4.0)
