+++ samples/javascript_typescript/11.qnamaker/cognitiveModels/smartLightFAQ.tsv
