+++ samples/javascript_typescript/11.qnamaker/deploymentScripts/msbotClone/bot.recipe
@@ -1,0 +1,38 @@
+{
+  "version": "1.0",
+  "resources": [
+    {
+      "type": "endpoint",
+      "id": "1",
+      "name": "development",
+      "url": "http://localhost:3978/api/messages"
+    },
+    {
+      "type": "qna",
+      "id": "34",
+      "name": "qnamaker-QNA"
+    },
+    {
+      "type": "endpoint",
+      "id": "2",
+      "name": "production",
+      "url": "https://your-bot-url.azurewebsites.net/api/messages"
+    },
+    {
+      "type": "abs",
+      "id": "3",
+      "name": "qna-abs"
+    },
+    {
+      "type": "appInsights",
+      "id": "4",
+      "name": "qna-insights"
+    },
+    {
+      "type": "blob",
+      "id": "5",
+      "name": "qna-blob",
+      "container": "botstatestore"
+    }
+  ]
+}
