+++ samples/javascript_typescript/11.qnamaker/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "qnamaker-ts",
+  "name": "qnamaker",
   "version": "1.0.0",
-  "description": "Bot Builder v4 QnA Maker with Applications Insights sample",
+  "description": "Bot Builder v4 QnA Maker sample",
   "main": "./lib/index.js",
   "scripts": {
     "build-sample": "tsc",
@@ -12,13 +12,15 @@
   "author": "Microsoft",
   "license": "MIT",
   "dependencies": {
-    "botbuilder": "^4.0.0-preview1.1135",
-    "botbuilder-ai": "^4.0.0-preview1.1135",
-    "botframework-config": "^4.0.0-preview1.1102",
+    "botbuilder": "^4.0.6",
+    "botbuilder-ai": "^4.0.6",
+    "botframework-config": "^4.0.6",
     "dotenv": "^6.0.0",
     "restify": "^6.4.0"
   },
   "devDependencies": {
-    "nodemon": "^1.18.4"
+    "nodemon": "^1.18.4",
+    "tslint": "^5.11.0",
+    "tslint-microsoft-contrib": "^5.2.1"
   }
 }
