+++ samples/javascript_typescript/11.qnamaker/qnamaker.bot
@@ -1,11 +1,10 @@
 {
     "name": "qnamaker",
-    "secretKey": "",
     "services": [
         {
             "type": "endpoint",
-            "name": "qnamaker",
-            "id": "http://localhost:3978/api/messages",
+            "name": "development",
+            "id": "1",
             "appId": "",
             "appPassword": "",
             "endpoint": "http://localhost:3978/api/messages"
@@ -21,4 +20,6 @@
             "ignoredValue": ""
         }
     ]
+    "padlock": "",
+    "version": "2.0"
 }
