+++ samples/javascript_typescript/11.qnamaker/tsconfig.json
